fix(page): requested changes
