+++ src/pages/Projects/report-pdf.tsx
@@ -2,82 +2,13 @@ import { Document, Page, Text, View } from '@react-pdf/renderer';
 import colors from '../../colors';
 import { Report } from '../../types/project-report';
 import { truncateText } from '../../utils/methods';
+import { numberToMonth, dateParser, statusColor } from './reportMethods';
 
 interface reportProps {
   data: Report;
   usingFilters: boolean;
   month: number;
   year: number;
-}
-
-function dateParser(date: Date): string {
-  const arr = date.toString().split('-');
-  const day = arr[2].substring(0, 2);
-  const month = arr[1];
-  const year = arr[0];
-  return `${day}-${month}-${year}`;
-}
-
-function numberToMonth(number: number): string | null {
-  switch (number) {
-    case 1:
-      return 'January';
-    case 2:
-      return 'February';
-    case 3:
-      return 'March';
-    case 4:
-      return 'April';
-    case 5:
-      return 'May';
-    case 6:
-      return 'June';
-    case 7:
-      return 'Jule';
-    case 8:
-      return 'August';
-    case 9:
-      return 'September';
-    case 10:
-      return 'October';
-    case 11:
-      return 'November';
-    case 12:
-      return 'December';
-    default:
-      return null;
-  }
-}
-
-function statusColor(status: string) {
-  status = status.toUpperCase();
-
-  switch (status) {
-    case 'ACCEPTED':
-      return colors.lightSuccess;
-    case 'NOT STARTED':
-      return colors.lightRed;
-    case 'IN PROGRESS':
-      return colors.warning;
-    case 'IN PROCESS':
-      return colors.warning;
-    case 'UNDER REVISION':
-      return colors.purple;
-    case 'DELAYED':
-      return colors.lightOrange;
-    case 'POSTPONED':
-      return colors.blue;
-    case 'DONE':
-      return colors.success;
-    case 'CANCELLED':
-      return colors.warning;
-    case 'IN QUOTATION':
-      return colors.darkBlue;
-    case 'AREA':
-      return colors.extra;
-    default:
-      return colors.null;
-  }
 }
 
 function infoComponent(title: string, value: string, style: string = '-') {
@@ -109,6 +40,7 @@ function infoComponent(title: string, value: string, style: string = '-') {
 const ProjectReportPDF = (props: reportProps) => {
   let tasks: number = -1;
   const totalTasks: number = Number(props.data?.statistics?.total) || 1;
+  const filterMonth = numberToMonth(props.month);
   const keyMap = new Map<string, string>([
     ['done', 'Done'],
     ['inprogress', 'In process'],
@@ -125,7 +57,7 @@ const ProjectReportPDF = (props: reportProps) => {
         <View style={{ color: 'black', textAlign: 'justify', margin: 30, gap: '20px' }}>
           {props.usingFilters === true && (
             <Text style={{ textAlign: 'right', fontSize: 12, fontFamily: 'Times-Bold' }}>
-              {`${numberToMonth(props.month) || ''} ${props.year}`}
+              {`${filterMonth !== 'Invalid month number' ? filterMonth : ''} ${props.year}`}
             </Text>
           )}
           <Text style={{ textAlign: 'center', fontSize: 26, fontFamily: 'Times-Bold' }}>
