+++ src/pages/Projects/report.tsx
@@ -23,27 +23,8 @@ import { axiosInstance } from '../../lib/axios/axios';
 import { Report } from '../../types/project-report';
 import { APIPath, BASE_API_URL, RequestMethods } from '../../utils/constants';
 import { truncateText } from '../../utils/methods';
+import { dateParser, filterteParser, capitalize } from './reportMethods';
 import ProjectReportPDF from './report-pdf';
-
-function dateParser(date: Date): string {
-  const arr = date.toString().split('-');
-  const day = arr[2].substring(0, 2);
-  const month = arr[1];
-  const year = arr[0];
-  return `${day}-${month}-${year}`;
-}
-
-function filterteParser(date: Date): string {
-  const arr = date.toISOString().split('-');
-  const day = arr[2].substring(0, 2);
-  const month = arr[1];
-  const year = arr[0];
-  return `${year}-${month}-${day}`;
-}
-
-function capitalize(data: string): string {
-  return data.charAt(0).toUpperCase() + data.substring(1).toLowerCase();
-}
 
 const ProjectReport: React.FC = () => {
   const { id } = useParams();
